+++ test/unit/feature/login.test.js
@@ -13,9 +13,17 @@ const main = async function() {
         expect(typeof response_de_login.data).to.equal("object");
         expect(typeof response_de_login.data.token_de_sesion_activo).to.equal("string");
         expect(typeof response_de_login.data.token_de_sesion_activo.length).to.equal("number");
+        expect(typeof response_de_login.data.autentificacion).to.equal("object");
+        expect(typeof response_de_login.data.autentificacion.sesion).to.equal("object");
+        expect(typeof response_de_login.data.autentificacion.usuario).to.equal("object");
+        expect(typeof response_de_login.data.autentificacion.grupos).to.equal("object");
+        expect(typeof response_de_login.data.autentificacion.grupos.length).to.equal("number");
+        expect(typeof response_de_login.data.autentificacion.permisos).to.equal("object");
+        expect(typeof response_de_login.data.autentificacion.permisos.length).to.equal("number");
+        console.log(response_de_login.data.autentificacion);
         log_verde("[✔] Test para login exitoso");
         pasar_test("login.test.js");
-        deployer.utilities.gestor_de_shutdown();
+        await deployer.utilities.gestor_de_shutdown();
     } catch(error) {
         log_rojo("[✘] Test para login fallido");
         fallar_test("login.test.js");